+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Current (in progress)
 
+- Use beta admin as default [#621](https://github.com/datagouv/udata-front/pull/621)
 - Fix homepage translations and links [#644](https://github.com/datagouv/udata-front/pull/644) [#648](https://github.com/datagouv/udata-front/pull/648)
 - Fix security tests after udata changes [#650](https://github.com/datagouv/udata-front/pull/650)
 - Add Bluesky link [#649](https://github.com/datagouv/udata-front/pull/649)